+++ .gitignore
@@ -1,3 +1,4 @@
 .env
 .flake8
 .txt
+*.pyc
