+++ Backend/requirements.txt
@@ -1,23 +1,28 @@
-aiosmtplib==2.0.2
+aiosmtplib==3.0.2
 annotated-types==0.7.0
 anyio==4.4.0
-autopep8==2.2.0
-bcrypt==4.1.3
-black==24.4.2
+asgiref==3.8.1
+autopep8==2.3.1
+bcrypt==4.2.0
+beanie==1.26.0
+black==24.8.0
 blinker==1.8.2
-certifi==2024.6.2
-cffi==1.16.0
+build==1.2.1
+certifi==2024.7.4
+cffi==1.17.0
 charset-normalizer==3.3.2
 click==8.1.7
-cryptography==42.0.8
+cryptography==43.0.0
+Django==5.1
+djangorestframework==3.15.2
 dnspython==2.6.1
 ecdsa==0.19.0
-email_validator==2.1.1
+email_validator==2.2.0
 emoji==2.12.1
-fastapi==0.111.0
-fastapi-cli==0.0.4
+fastapi==0.112.1
+fastapi-cli==0.0.5
 fastapi-mail==1.4.1
-flake8==7.0.0
+flake8==7.1.1
 Flask==3.0.3
 greenlet==3.0.3
 h11==0.14.0
@@ -25,62 +30,64 @@ httpcore==1.0.5
 httptools==0.6.1
 httpx==0.27.0
 idna==3.7
-inflect==7.2.1
+inflect==7.3.1
 iniconfig==2.0.0
 isort==5.13.2
 itsdangerous==2.2.0
 Jinja2==3.1.4
+lazy-model==0.2.0
 markdown-it-py==3.0.0
 MarkupSafe==2.1.5
 mccabe==0.7.0
 mdurl==0.1.2
-more-itertools==10.3.0
+more-itertools==10.4.0
+motor==3.5.1
 mypy-extensions==1.0.0
-orjson==3.10.5
+orjson==3.10.7
 packaging==24.1
 passlib==1.7.4
 pathspec==0.12.1
 pip-review==1.3.0
+pip-tools==7.4.1
 platformdirs==4.2.2
 pluggy==1.5.0
 pyasn1==0.6.0
-pycodestyle==2.11.1
+pycodestyle==2.12.1
 pycparser==2.22
-pydantic==2.7.4
-pydantic-settings==2.3.3
-pydantic_core==2.18.4
+pydantic==2.8.2
+pydantic-settings==2.4.0
+pydantic_core==2.20.1
 pyfiglet==1.0.2
 pyflakes==3.2.0
 Pygments==2.18.0
-pytest==8.2.2
+PyJWT==2.9.0
+pymongo==4.8.0
+pyproject_hooks==1.1.0
+pytest==8.3.2
 python-decouple==3.8
 python-dotenv==1.0.1
 python-jose==3.3.0
 python-multipart==0.0.9
-PyYAML==6.0.1
+PyYAML==6.0.2
 requests==2.32.3
 rich==13.7.1
 rsa==4.9
 shellingham==1.5.4
 six==1.16.0
 sniffio==1.3.1
-SQLAlchemy==2.0.30
+SQLAlchemy==2.0.32
 SQLAlchemy-Utils==0.41.2
-sqlmodel==0.0.19
-starlette==0.37.2
+sqlmodel==0.0.21
+sqlparse==0.5.1
+starlette==0.38.2
+toml==0.10.2
 typeguard==4.3.0
 typer==0.12.3
 typing_extensions==4.12.2
 ujson==5.10.0
-urllib3==2.2.1
-uvicorn==0.30.1
-uvloop==0.19.0
-watchfiles==0.22.0
+urllib3==2.2.2
+uvicorn==0.30.6
+uvloop==0.20.0
+watchfiles==0.23.0
 websockets==12.0
 Werkzeug==3.0.3
-pymongo==4.8.0
-dnspython>=1.16.0,<3.0.0
-beanie=1.26.0 
-lazy-model=0.2.0 
-motor=3.5.1 
-toml=0.10.2
